`._plrease_do_not_use`というモジュールを切り、`_Reserved`をそこに入れる
